requirements.txt
Pillow==11.0.0
ttkbootstrap==1.10.1
rich==13.9.4
customtkinter==5.2.2
